Update core for CUDA nvcc
* nvcc says it supports C++20, but there's still a lot of bugs from what I can see. I tried to report them, but their system is way to annoying. The main bugs are related `[[no_unique_address]]` and `if constexpr` with variadics and requires expression. I had to write significant changes to make it compile (and give correct results) with nvcc.

* Update `Span` API to make it closer to `std::span`.

* Add `Accessor::stride(int)` and `guts::are_accessors_const`. Few other small changes/fixes.

* Remove clang warning in Blas.cpp and conversion warnings

+++ tests/noa/cpu/TestCPUReduceEwise.cpp
@@ -19,9 +19,9 @@ TEST_CASE("cpu::reduce_ewise") {
         const auto shape = Shape4<i64>{1, 50, 1, 100};
         const auto elements = shape.elements();
 
-        const auto buffer = std::make_unique<f32[]>(elements);
+        const auto buffer = std::make_unique<f64[]>(elements);
         std::fill_n(buffer.get(), elements, 1);
-        auto input = noa::make_tuple(AccessorI64<f32, 4>(buffer.get(), shape.strides()));
+        auto input = noa::make_tuple(AccessorI64<f64, 4>(buffer.get(), shape.strides()));
         auto init = noa::make_tuple(AccessorValue<f64>(0.));
         auto output = noa::make_tuple(AccessorValue<f64>(0.));
